Tech Exercise
Write a program that allows for an integer array to be passed in and
will then output all of the pairs that sum up to 10.  Please provide a
solution that allows for 1) output all pairs (includes duplicates and
the reversed ordered pairs), 2) output unique pairs only once (removes
the duplicates but includes the reversed ordered pairs), and 3) output
the same combo pair only once (removes the reversed ordered pairs).